refactor out lower level USB detection - will make testing mocks easier
